+++ eternity/source/d_net.h
@@ -124,37 +124,37 @@ typedef struct doomdata_s doomdata_t;
 struct doomcom_s
 {
     // Supposed to be DOOMCOM_ID?
-    long                id;
+    int32_t             id;
     
     // DOOM executes an int to execute commands.
-    short               intnum;         
+    int16_t             intnum;         
     // Communication between DOOM and the driver.
     // Is CMD_SEND or CMD_GET.
-    short               command;
+    int16_t             command;
     // Is dest for send, set by get (-1 = no packet).
-    short               remotenode;
+    int16_t             remotenode;
     
     // Number of bytes in doomdata to be sent
-    short               datalength;
+    int16_t             datalength;
 
     // Info common to all nodes.
     // Console is allways node 0.
-    short               numnodes;
+    int16_t             numnodes;
     // Flag: 1 = no duplication, 2-5 = dup for slow nets.
-    short               ticdup;
+    int16_t             ticdup;
     // Flag: 1 = send a backup tic in every packet.
-    short               extratics;
+    int16_t             extratics;
     // Flag: 1 = deathmatch.
-    short               deathmatch;
+    int16_t             deathmatch;
     // Flag: -1 = new game, 0-5 = load savegame
-    short               savegame;
-    short               episode;        // 1-3
-    short               map;            // 1-9
-    short               skill;          // 1-5
+    int16_t             savegame;
+    int16_t             episode;        // 1-3
+    int16_t             map;            // 1-9
+    int16_t             skill;          // 1-5
 
     // Info specific to this node.
-    short               consoleplayer;
-    short               numplayers;
+    int16_t             consoleplayer;
+    int16_t             numplayers;
     
     // These are related to the 3-display mode,
     //  in which two drones looking left and right
@@ -162,9 +162,9 @@ struct doomcom_s
     //  on two additional computers.
     // Probably not operational anymore.
     // 1 = left, 0 = center, -1 = right
-    short               angleoffset;
+    int16_t             angleoffset;
     // 1 = drone
-    short               drone;          
+    int16_t             drone;          
 
     // The packet data to be sent.
     doomdata_t          data;
