+++ eternity/source/e_lib.h
@@ -61,7 +61,7 @@ const char *E_BuildDefaultFn(const char *filename);
 // misc utilities
 int E_EnableNumForName(const char *name, E_Enable_t *enables);
 int E_StrToNumLinear(const char **strings, int numstrings, const char *value);
-long E_ParseFlags(const char *str, dehflagset_t *flagset);
+int E_ParseFlags(const char *str, dehflagset_t *flagset);
 char *E_ExtractPrefix(char *value, char *prefixbuf, int buflen);
 
 // keywords
